Rename EndermanTeleportEvent to EntityTeleportEvent.java, correct javadocs per @EvilSeph
